Added ips to django's ALLOWED_HOSTS
Acceing the server from the host machine triggers ERR_CONNECTION_REFUSED since the VM binds to 127.0.0.1, which is loopback.
This commit makes sure that we can bind the server on the guest machine to 0.0.0.0 so that all interfaces can access it.
In addition, the local host ip was added since we've modified the default specification that includes it (and we still want to inculde it).