Cleaned rebase/editor conflicts headers.
This branch started from merging a development branch (11galpid/dev-userlogin) that later was reduced, in order to simplify it's PR.
Therefore, while continuing development on this branch, it was removed as well at some point. This caused some turbulence in the existing
state of the common files between the branches. This commit is CLEANUP commit.