Added the Stream Model to support querying previous or scheduled streams.
The table name is set to 'streams' (to avoid a table name such as 'main_app_stream') and ordering is done by the 'airs_on' and 'author' field to support querying the most recent streams.

Note that the Model uses Django's default User model. This model will be extended as part of a different task.

In addition, updated the streams table specification in the README file.

Lastly, the migration file for the Model was added under the migrations folder.